Moved the UI text field to a separate file
