+++ src/common/CMakeLists.txt
@@ -64,6 +64,8 @@ add_library(CommonFiles STATIC
     ui/MI_ImageSelectField.h
     ui/MI_SliderField.cpp
     ui/MI_SliderField.h
+    ui/MI_TextField.h
+    ui/MI_TextField.cpp
 )
 
 if (USE_PNG_SAVE AND NOT USE_SDL2_LIBS)
