+++ src/common/uicontrol.h
@@ -272,80 +272,4 @@ protected:
     std::function<void()> onPressFn;
 };
 
-class MI_TextField : public UI_Control
-{
-public:
-
-    MI_TextField(gfxSprite * nspr, short x, short y, const char * name, short width, short indent);
-    virtual ~MI_TextField();
-
-    void SetTitle(char * name);
-
-    //Gets the values of the currently selected item
-    char * GetValue() {
-        return szValue;
-    }
-
-    //Called when user selects this control to change it's value
-    MenuCodeEnum Modify(bool modify);
-
-    void Clear() {
-        szValue[0] = 0;
-        iCursorIndex = 0;
-        iNumChars = 1;
-    }
-
-    //Updates animations or other events every frame
-    void Update();
-
-    //Draws every frame
-    virtual void Draw();
-
-    //Sends player input to control on every frame
-    virtual MenuCodeEnum SendInput(CPlayerInput * playerInput);
-
-    //When the item is changed, this code will be returned from SendInput()
-    void SetItemChangedCode(MenuCodeEnum code) {
-        mcItemChangedCode = code;
-    }
-    void SetControlSelectedCode(MenuCodeEnum code) {
-        mcControlSelectedCode = code;
-    }
-
-    //Set where the data of this control is written to (some member of game_values probably)
-    void SetData(char * value, short maxchars);
-
-    MenuCodeEnum MouseClick(short iMouseX, short iMouseY);
-
-    void Refresh();
-
-    void SetDisallowedChars(const char * chars);
-
-protected:
-
-    void UpdateCursor();
-
-    short iCursorIndex;
-    short iNumChars, iMaxChars;
-
-    gfxSprite * spr;
-    char * szName;
-
-    char * szValue;
-
-    short iWidth, iIndent;
-
-    MI_Image * miModifyCursor;
-
-    MenuCodeEnum mcItemChangedCode;
-    MenuCodeEnum mcControlSelectedCode;
-
-    short iAdjustmentY;
-    char * szTempValue;
-
-    short iStringWidth, iAllowedWidth;
-
-    char szDisallowedChars[32];
-};
-
 #endif // UICONTROL_H
