+++ src/smw/menu/network/NetLobbyMenu.h
@@ -5,6 +5,8 @@
 #include "uicontrol.h"
 #include "uicustomcontrol.h"
 
+class MI_TextField;
+
 /*
     Description.
 */
