+++ src/smw/menu/network/NetRoomMenu.h
@@ -6,6 +6,8 @@
 #include "uicustomcontrol.h"
 #include "ui/MI_NetRoomTeamSelect.h"
 
+class MI_TextField;
+
 /*
     Description.
 */
